Package resolver statically typed
