+++ src/main/kotlin/com/cognifide/gradle/aem/api/AemExtension.kt
@@ -4,9 +4,8 @@ import com.cognifide.gradle.aem.instance.Instance
 import com.cognifide.gradle.aem.instance.InstancePlugin
 import com.cognifide.gradle.aem.instance.InstanceSync
 import com.cognifide.gradle.aem.pkg.PackagePlugin
-import groovy.lang.Closure
+import org.gradle.api.Action
 import org.gradle.api.Project
-import org.gradle.util.ConfigureUtil
 
 /**
  * Main place for providing build script DSL capabilities in case of AEM.
@@ -36,36 +35,20 @@ open class AemExtension(@Transient private val project: Project) {
         instances.parallelStream().forEach { it.sync(synchronizer) }
     }
 
-    fun config(configurer: AemConfig.() -> Unit) {
-        config.apply(configurer)
+    fun config(configurer: Action<AemConfig>) {
+        configurer.execute(config)
     }
 
-    fun config(closure: Closure<*>) {
-        config { ConfigureUtil.configure(closure, this) }
+    fun bundle(configurer: Action<AemBundle>) {
+        configurer.execute(bundle)
     }
 
-    fun bundle(configurer: AemBundle.() -> Unit) {
-        bundle.apply(configurer)
+    fun notifier(configurer: Action<AemNotifier>) {
+        configurer.execute(notifier)
     }
 
-    fun bundle(closure: Closure<*>) {
-        bundle { ConfigureUtil.configure(closure, this) }
-    }
-
-    fun notifier(configurer: AemNotifier.() -> Unit) {
-        notifier.apply(configurer)
-    }
-
-    fun notifier(closure: Closure<*>) {
-        notifier { ConfigureUtil.configure(closure, this) }
-    }
-
-    fun tasks(configurer: AemTaskFactory.() -> Unit) {
-        tasks.apply(configurer)
-    }
-
-    fun tasks(closure: Closure<*>) {
-        tasks { ConfigureUtil.configure(closure, this) }
+    fun tasks(configurer: Action<AemTaskFactory>) {
+        configurer.execute(tasks)
     }
 
     companion object {
