+++ src/main/kotlin/com/cognifide/gradle/aem/base/vlt/CleanTask.kt
@@ -2,10 +2,9 @@ package com.cognifide.gradle.aem.base.vlt
 
 import com.cognifide.gradle.aem.api.AemDefaultTask
 import com.cognifide.gradle.aem.internal.Formats
-import groovy.lang.Closure
+import org.gradle.api.Action
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
-import org.gradle.util.ConfigureUtil
 
 open class CleanTask : AemDefaultTask() {
 
@@ -24,8 +23,8 @@ open class CleanTask : AemDefaultTask() {
         notifier.default("Cleaned JCR content", "Directory: ${Formats.rootProjectPath(config.contentPath, project)}")
     }
 
-    fun settings(configurer: Closure<*>) {
-        ConfigureUtil.configure(configurer, runner.cleaner)
+    fun settings(configurer: Action<VltCleaner>) {
+        runner.cleaner.apply { configurer.execute(this) }
     }
 
     companion object {
